[Feature] Range Practice | Warm up before grinding (#29)
This is a massive feature that I really wanted to introduce. It helps you to improve your game through a quiz challenge. You can select the spot based on all available ranges and practice it.

On this feature you will be randomly asked your action on a specific spot. You can decide what to do and the app will show if your move is correct (by correct it means that according to your defined range, you should or should not play the hand).