Refactor: Update Gradle version to 8.11.1
This commit updates the Gradle version to 8.11.1. It also includes changes to the Gradle wrapper scripts and properties file to reflect this update.

Additionally, the Android Gradle Plugin version is updated to 8.10.1, Google Services plugin to 4.4.2, Firebase Crashlytics Gradle plugin to 3.0.4, and Navigation Safe Args Gradle plugin to 2.9.0.

The application version is bumped to 1.7.2.